+++ intern/remote/boss/remote.py
@@ -1,4 +1,4 @@
-﻿"""
+"""
 # Copyright 2016 The Johns Hopkins University Applied Physics Laboratory
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -71,6 +71,14 @@ class BossRemote(Remote):
         self._init_project_service(version)
         self._init_metadata_service(version)
         self._init_volume_service(version)
+
+    def __repr__(self):
+        """
+        Stringify the Remote.
+
+        Returns a representation of the BossRemote that lists the host.
+        """
+        return "<intern.remote.BossRemote [" + self._config['Default']['host'] + "]>"
 
     def _init_project_service(self, version):
         """
@@ -851,6 +859,22 @@ class BossRemote(Remote):
         self.metadata_service.set_auth(self._token_metadata)
         self.metadata_service.delete(resource, keys)
 
+    def parse_bossURI(self, uri): # type: (str) -> Resource
+        """
+        Parse a bossDB URI and handle malform errors.
+
+        Arguments:
+            uri (str): URI of the form bossdb://<collection>/<experiment>/<channel>
+
+        Returns:
+            Resource
+
+        """
+        t = uri.split("://")[1].split("/")
+        if len(t) is 3:
+            return self.get_channel(t[2], t[0], t[1])
+        raise ValueError("Cannot parse URI " + uri + ".")
+
     def get_cutout(self, resource, resolution, x_range, y_range, z_range, time_range=None, id_list=[], no_cache=None, access_mode=CacheMode.no_cache, **kwargs):
             """Get a cutout from the volume service.
 
@@ -860,19 +884,24 @@ class BossRemote(Remote):
             requester.
 
             Args:
-                resource (intern.resource.boss.resource.ChannelResource): Channel or layer resource.
+                resource (intern.resource.boss.resource.ChannelResource | str): Channel or layer Resource. If a 
+                    string is provided instead, BossRemote.parse_bossURI is called instead on a URI-formatted 
+                    string of the form `bossdb://collection/experiment/channel`.
                 resolution (int): 0 indicates native resolution.
                 x_range (list[int]): x range such as [10, 20] which means x>=10 and x<20.
                 y_range (list[int]): y range such as [10, 20] which means y>=10 and y<20.
                 z_range (list[int]): z range such as [10, 20] which means z>=10 and z<20.
                 time_range (optional [list[int]]): time range such as [30, 40] which means t>=30 and t<40.
                 id_list (optional [list[int]]): list of object ids to filter the cutout by.
+                no_cache (optional [boolean or None]): Deprecated way to specify the use of cache to be True or False.
+                    access_mode should be used instead
                 access_mode (optional [Enum]): Identifies one of three cache access options:
                     cache = Will check both cache and for dirty keys
                     no_cache = Will skip cache check but check for dirty keys
                     raw = Will skip both the cache and dirty keys check
 
                 TODO: Add mode to documentation
+
 
             Returns:
                 (numpy.array): A 3D or 4D (time) numpy matrix in (time)ZYX order.
@@ -892,3 +921,50 @@ class BossRemote(Remote):
             elif no_cache == False:
                 access_mode=CacheMode.cache
             return self._volume.get_cutout(resource, resolution, x_range, y_range, z_range, time_range, id_list, access_mode, **kwargs)
+
+    def get_experiment(self, coll_name, exp_name):
+        """
+        Convenience method that gets experiment resource.
+        
+        Args:
+            coll_name (str): Collection name
+            exp_name (str): Experiment name
+
+        Returns:
+            (ExperimentResource)
+        """
+        exp = ExperimentResource(exp_name, coll_name)
+        return self.get_project(exp)
+
+    def get_coordinate_frame(self, name):
+        """
+        Convenience method that gets coordinate frame resource
+
+        Args:
+            name (str): Name of the coordinate frame
+        
+        Returns:
+            (CoordinateFrameResource)
+        """
+        cf = CoordinateFrameResource(name)
+        return self.get_project(cf)
+
+    def get_neuroglancer_link(self, resource, resolution, x_range, y_range, z_range, **kwargs):
+            """
+            Get a neuroglancer link of the cutout specified from the host specified in the remote configuration step. 
+
+            Args:
+                resource (intern.resource.Resource): Resource compatible with cutout operations.
+                resolution (int): 0 indicates native resolution.
+                x_range (list[int]): x range such as [10, 20] which means x>=10 and x<20.
+                y_range (list[int]): y range such as [10, 20] which means y>=10 and y<20.
+                z_range (list[int]): z range such as [10, 20] which means z>=10 and z<20.
+
+            Returns:
+                (string): Return neuroglancer link.
+
+            Raises:
+                RuntimeError when given invalid resource.
+                Other exceptions may be raised depending on the volume service's implementation.
+            """
+            return self._volume.get_neuroglancer_link(resource, resolution, x_range, y_range, z_range, **kwargs)