+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 ____________
+## v0.9.11
+ Convenience functions added: 
+  * `BossRemote#get_experiment()`
+  * `BossRemote#get_coordinate_frame()`
+  * `BossRemote#get_neuroglancer_link()` get a neuroglancer URL from cutout arguments 
+* `create_cutout` will now chunk large cutouts into smaller components to eliminate HTTP size limitation and timeout errors.
+
 ## v0.9.10
 *  Intern now forwards no_cache option when breaking apart large cutouts.